+++ Runtime/Core/DisplayController.cs
@@ -1,7 +1,10 @@
 using UnityEngine;
+using UnityEngine.Rendering;
 
 namespace Pixygon.Micro {
     public class DisplayController : MonoBehaviour {
         public Camera _camera;
+        public Volume _volume;
+        public VolumeProfile _defaultVolume;
     }
 }